Create Cleanup-mailboxes.ps1
Simple Script for removing Active Directory Accounts and associated mailbox.  Includes a parameter to also remove the HomeDrive then sends an audit report by mail.  I have left the write-host lines of code in so that you can see the output in EAC or Powershell 
session when the process completes to compare with the results of the HTML report.